+++ sonos-control-api/src/main/java/de/kalass/sonoscontrol/api/model/renderingcontrol/Bass.java
@@ -4,13 +4,13 @@ import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 
 public final class Bass {
-    private final Integer _value;
+    private final Long _value;
 
-    private Bass(final Integer value) {
+    private Bass(final Long value) {
         _value = Preconditions.checkNotNull(value);
     }
 
-    public Integer asInteger() {
+    public Long asLong() {
         return _value;
     }
 
@@ -32,7 +32,7 @@ public final class Bass {
         return Objects.toStringHelper(this).add("value", _value).toString();
     }
 
-    public static Bass valueOf(Integer value) {
+    public static Bass valueOf(Long value) {
         return value == null ? null : new Bass(value);
     }
 }
